+++ HISTORY.rst
@@ -6,6 +6,8 @@ History
 Future minor version (release soon)
 -----------------------------------
 
+* Remove import of matplotlib backend QtAgg and related error message, except for sarvey_plot.
+
 1.3.0 (2026-02-23)
 ------------------